requirements.txt
cryptography~=3.3.2
PyYAML~=5.3.1
SQLAlchemy~=1.4.22
pandas~=1.3.1
numpy~=1.21.1
ccxt~=1.53.97
statsmodels~=0.12.2
PyQt5~=5.15.4
Quandl~=3.6.1
yfinance~=0.1.63
investpy~=1.0.7
requests~=2.25.1
